+++ app/controllers/api/v1/registered_organisations_schemes_controller.rb
@@ -1,15 +1,20 @@
 module Api
   module V1
     class RegisteredOrganisationsSchemesController < ActionController::API
-      include Authorize::User
+      include Authorize::AuthorizationMethods
       rescue_from ApiValidations::ApiError, with: :return_error_code
-      before_action :validate_ccs_org_user_or_api_key
+      before_action :validate_access_users_or_api_key
       before_action :validate_params
 
-      def search_organisation
+      def search_organisation 
         params[:ccs_org_id] = search_organisation_by_scheme if @scheme_id
 
-        result = Common::RegisteredOrganisationResponse.new(params[:ccs_org_id], hidden: false).response_payload if params[:ccs_org_id]
+        if validate_service_eligibility_or_ccs_admin_user
+          result = Common::RegisteredOrganisationResponse.new(params[:ccs_org_id], hidden: true).response_payload if params[:ccs_org_id]
+        else
+          result = Common::RegisteredOrganisationResponse.new(params[:ccs_org_id], hidden: false).response_payload if params[:ccs_org_id]
+        end
+
         if result.present?
           render json: build_response(result), status: :ok
         else
