+++ app/controllers/api/v1/remove_organisations_additional_identifier_controller.rb
@@ -2,7 +2,7 @@ module Api
   module V1
     class RemoveOrganisationsAdditionalIdentifierController < ActionController::API
       include Authorize::Token
-      include Authorize::User
+      include Authorize::AuthorizationMethods
       rescue_from ApiValidations::ApiError, with: :return_error_code
       before_action :validate_api_key
       before_action :validate_user
