+++ app/controllers/api/v1/update_organisations_controller.rb
@@ -2,7 +2,7 @@ module Api
   module V1
     class UpdateOrganisationsController < ActionController::API
       include Authorize::Token
-      include Authorize::User
+      include Authorize::AuthorizationMethods
       rescue_from ApiValidations::ApiError, with: :return_error_code
       before_action :validate_api_key
       before_action :validate_user
