+++ app/services/authorize/user.rb
@@ -36,6 +36,22 @@ module Authorize
       ApiValidations::ApiErrorValidationResponse.new(:user_access_unauthorized) unless decoded_token[0]['roles'].include?(ENV['ACCESS_DATA_MIGRATION'])
     end
 
+    def validate_service_eligibility_user
+      decoded_token = validate_and_decode_token
+      ApiValidations::ApiErrorValidationResponse.new(:user_access_unauthorized) unless decoded_token[0]['roles'].include?(ENV['ACCESS_MANAGE_SUBSCRIPTIONS'])
+    end
+
+    def validate_service_eligibility_or_ccs_admin_user
+      decoded_token = validate_and_decode_token
+      if decoded_token[0]['roles'].include?(ENV['ACCESS_MANAGE_SUBSCRIPTIONS'])
+        true
+      elsif decoded_token[0]['roles'].include?(ENV['ACCESS_CCS_ADMIN'])
+        false
+      else
+        ApiValidations::ApiErrorValidationResponse.new(:user_access_unauthorized)
+      end
+    end
+
     def validate_access_token
       decoded_token = validate_and_decode_token
       validate_token = SecurityService::Auth.new(decoded_token[0]['aud'], Common::ApiHelper.bearer_token(request.headers)).sec_api_validate_token
@@ -65,58 +81,6 @@ module Authorize
       return false if Client.find_by(api_key: api_token.to_s)&.id.blank?
 
       true
-    end
-
-    def validate_no_role
-      validate_client_id
-      validate_user_access_token
-      validate_access_token
-      validate_ccs_org_id
-    end
-
-    def validate_user
-      validate_client_id
-      validate_user_access_token
-      validate_access_token
-      validate_organisation_user
-      validate_ccs_org_id
-    end
-
-    def validate_ccs_admin
-      validate_client_id
-      validate_user_access_token
-      validate_access_token
-      validate_ccs_admin_user
-      validate_ccs_org_id
-    end
-
-    def validate_integrating_service_user
-      return if validate_integration_token
-
-      validate_client_id
-      validate_user_access_token
-      validate_access_token
-      validate_data_migration_user
-    end
-
-    def validate_ccs_org_user_or_api_key
-      return if validate_api_token
-
-      validate_client_id
-      validate_user_access_token
-      validate_access_token
-      validate_organisation_user
-      validate_ccs_org_id
-    end
-
-    def validate_ccs_admin_or_api_key
-      return if validate_api_token
-
-      validate_client_id
-      validate_user_access_token
-      validate_access_token
-      validate_ccs_admin_user
-      validate_ccs_org_id
     end
   end
 end
